Find members of the Security group types
The task was to find those users who are members of group with type 'security'
Create a GlideRecord object with the table name used in your organization's instance. First query for the group.type Contains the sys_id of the group type. Then, query for active users and get their sys_ids into the array groupNames. Because the script returns an array of user's sys_id, we might have to use this with the sys_user table to match the results. Other settings include APIName: global.SecurityGroupUsers, Client callable: true, Active: true, Accessible from: This Application Only, Application: Global.