How does this script help?
The requirement was to find the users who are a members of at least one group, and those members who are not a member of the BASE ITIL Group
#ServiceNow #The_Bharatha

The way script works is, the first GlideRecord will first query for all the members in the BASE ITIL Group, and then use the second GlideRecord will query for all the members across all the other groups. The array of User's sys_ids from both Base ITIL group and other groups are pushed into two different arrays. At the end of the script, the ArrayUtil will take a difference between both the arrays, the result will be am array of the User's sys_ids who are not a member of Base ITIL Group.